+++ assets/modules/main.mjs
@@ -1,45 +1,33 @@
 import { CanvasDrawer } from "./canvasDrawer.mjs";
-import { CellType } from "./cellType.mjs";
+import { Empire } from "./empire.mjs";
 import { FadingLayeredPerlinGenerator } from "./worldGenerators/fadingLayeredPerlinGenerator.mjs";
 
 
 
 const CANVAS = document.getElementById("canvas");
+const STEPPER = document.getElementById("step");
 
 
 
 let world = FadingLayeredPerlinGenerator.create({
-    width: 40,
-    height: 40,
-    zoom: 20,
-    seed: 1337
+    width: 100,
+    height: 100,
+    zoom: 20
 });
 
 
 
-let center = world.getCell(0, 39)
-
-center.type = CellType.DBG_RED;
-
-let neighbors = center.getNeighbors()
-
-let types = [
-    CellType.DBG_BLUE,
-    CellType.DBG_BROWN,
-    CellType.DBG_CYAN,
-    CellType.DBG_GREEN,
-    CellType.DBG_MAGENTA,
-    CellType.DBG_ORANGE,
-    CellType.DBG_PINK,
-    CellType.DBG_YELLOW
-]
-
-for (let i = 0; i < neighbors.length; i++) {
-    if (neighbors[i]) {
-        neighbors[i].type = types[i]
-    }
-}
+world.newEmpire("empireA", "#ff0000ff", [world.getCell(50, 50)])
+world.newEmpire("empireB", "#0000ffff", [world.getCell(30, 50)])
 
 
 
-CanvasDrawer.draw(world, CANVAS, 100)
+CanvasDrawer.draw(world, CANVAS, 40);
+
+
+
+const LOOP = setInterval(function () {
+    world.step();
+
+    CanvasDrawer.draw(world, CANVAS, 40);
+});