+++ assets/modules/worldGenerators/fadingLayeredPerlinGenerator.mjs
@@ -21,7 +21,6 @@ export class FadingLayeredPerlinGenerator extends Generator {
         noise.seed(seed);
 
         let world = new World([]);
-        let cells = [];
 
         for (let x = 0; x < width; x++) {
             for (let y = 0; y < height; y++) {
@@ -45,11 +44,9 @@ export class FadingLayeredPerlinGenerator extends Generator {
                     type = CellType.MOUNTAIN
                 }
 
-                cells.push(new Cell(x, y, type, world));
+                new Cell(x, y, type, world);
             }
         }
-
-        world.cells = cells;
 
         return world;
     }